Updating the upload picture
