+++ src/main/java/com/example/vms/controller/ImageController.java
@@ -1,0 +1,33 @@
+package com.example.vms.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.multipart.MultipartFile;
+
+import com.example.vms.service.ImageService;
+import com.example.vms.util.BasicResponse;
+
+@RestController
+public class ImageController {
+	
+	private final ImageService imageService;
+
+	public ImageController(ImageService imageService) {
+		super();
+		this.imageService = imageService;
+	}
+	
+	@PostMapping("/upload-picture")
+	public ResponseEntity<BasicResponse> uploadPicture(@RequestParam ("userId") int userId, 
+							@RequestParam ("file") MultipartFile file){
+		imageService.uploadPicture(userId, file);
+		return ResponseEntity.status(HttpStatus.CREATED)
+				.body(BasicResponse.create(HttpStatus.CREATED.value(), "Profile Picture is Updated"));
+	}
+	
+	
+	
+}
