+++ src/main/java/com/example/vms/controller/UserController.java
@@ -4,19 +4,16 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.multipart.MultipartFile;
 
 import com.example.vms.entity.User;
 import com.example.vms.service.UserService;
-import com.example.vms.util.BasicResponse;
 import com.example.vms.util.ResponseStructure;
 
 @RestController
 public class UserController {
 	
-	private UserService userService;
+	private final UserService userService;
 
 	public UserController(UserService userService) {
 		super();
@@ -30,13 +27,5 @@ public class UserController {
 				.body(ResponseStructure.create(HttpStatus.CREATED.value(), "User Created", user));
 	}
 	
-	@PostMapping("/upload-picture")
-	public ResponseEntity<BasicResponse> uploadPicture(@RequestParam ("userId") int userId, 
-							@RequestParam ("file") MultipartFile file){
-		userService.uploadPicture(userId, file);
-		return ResponseEntity.status(HttpStatus.CREATED)
-				.body(BasicResponse.create(HttpStatus.CREATED.value(), "User Profile Picture is Updated"));
-		
-	}
-
+	
 }
